+++ src/main/java/notification/Control/ClearNotifInputBoundary.java
@@ -1,4 +1,4 @@
-package notification.clear_notif;
+package notification.Control;
 //use case layer
 public interface ClearNotifInputBoundary {
     void clearNotif();
