+++ src/main/java/notification/Control/ShowNotifInputBoundary.java
@@ -1,4 +1,4 @@
-package notification.show_notif;
+package notification.Control;
 //use case layer
 
 public interface ShowNotifInputBoundary {
