+++ src/main/java/notification/Present/ClearNotifOutputBoundary.java
@@ -1,4 +1,4 @@
-package notification.clear_notif;
+package notification.Present;
 //use case ayer
 public interface ClearNotifOutputBoundary {
     void prepareDisplayView();
