+++ src/main/java/notification/Present/ClearNotifPresenter.java
@@ -1,8 +1,8 @@
-package notification.clear_notif;
+package notification.Present;
 
-import ui.NotificationUI;
+import notification.NotificationUI;
 
-public class ClearNotifPresenter implements ClearNotifOutputBoundary{
+public class ClearNotifPresenter implements ClearNotifOutputBoundary {
     NotificationUI notifUI;
 
     public ClearNotifPresenter(){}
