+++ src/main/java/notification/Present/ShowNotifOutputBoundary.java
@@ -1,4 +1,7 @@
-package notification.show_notif;
+package notification.Present;
+
+import notification.Present.NotifResponseModel;
+
 //use case layer
 public interface ShowNotifOutputBoundary {
     void prepareDisplayView(NotifResponseModel response);
